Fixes #9657 - fix typo in test case - which then failed - and then fix the failure !strict
git-svn-id: http://svn.dojotoolkit.org/src/dijit/trunk@19904 560b804f-0ae3-0310-86f3-f6aa0a117693

+++ _Widget.js
@@ -548,7 +548,7 @@ dojo.declare("dijit._Widget", null, {
 
 		// destroy widgets created as part of template, etc.
 		dfe(this._supportingWidgets||[], function(w){ 
-			if(w.destroyRecursive){
+			if(w.isContainer && w.destroyRecursive){
 				w.destroyRecursive();
 			}else if(w.destroy){
 				w.destroy();
